Cast wasn't meant to be removed

+++ runelite-client/src/main/java/net/runelite/client/plugins/runenergy/RunEnergyPlugin.java
@@ -188,7 +188,7 @@ public class RunEnergyPlugin extends Plugin
 		// Return the text
 		if (inSeconds)
 		{
-			return Math.floor(secondsLeft) + "s";
+			return (int) Math.floor(secondsLeft) + "s";
 		}
 		else
 		{
